% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaanalysis.R
\name{.bayesianUpdate}
\alias{.bayesianUpdate}
\title{Perform Bayesian normal-normal update using neighbor prior}
\usage{
.bayesianUpdate(ests, ses, nn, patch_names)
}
\arguments{
\item{ests}{Genes x patches matrix of estimates.}

\item{ses}{Genes x patches matrix of standard errors.}

\item{nn}{Integer matrix (patches x k) of neighbor indices.}

\item{patch_names}{Character vector of patch names (column order).}
}
\value{
List with $ests, $ses, $pvals matrices (genes x patches).
}
\description{
Perform Bayesian normal-normal update using neighbor prior
}
