% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatches.R
\name{.buildContiguityGraph}
\alias{.buildContiguityGraph}
\title{Build contiguity graph using symmetric k-NN}
\usage{
.buildContiguityGraph(xy, k = 10)
}
\arguments{
\item{xy}{n x 2 coordinate matrix.}

\item{k}{Number of neighbors for k-NN. Default 10.}
}
\value{
Sparse binary adjacency matrix (n x n).
}
\description{
Two cells are connected if either is among the other's k nearest neighbors.
Much faster than Delaunay triangulation for large n.
}
