% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaanalysis.R
\name{.computeEnrichment}
\alias{.computeEnrichment}
\title{Compute enrichment scores for a set of patches relative to overall}
\usage{
.computeEnrichment(sg_patches, patch_meta)
}
\arguments{
\item{sg_patches}{Character vector of patch names in the subgroup.}

\item{patch_meta}{Output of .patchMetaSummary.}
}
\value{
Single-row data frame of enrichment columns.
}
\description{
Compute enrichment scores for a set of patches relative to overall
}
