% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{.dropOrphans}
\alias{.dropOrphans}
\title{Drop orphan cells that aren't contiguous with the main body of their patch}
\usage{
.dropOrphans(patch, spatial_nn)
}
\arguments{
\item{patch}{Character vector of patch assignments.}

\item{spatial_nn}{Sparse adjacency matrix (n x n).}
}
\value{
Updated patch vector with orphans set to NA.
}
\description{
Drop orphan cells that aren't contiguous with the main body of their patch
}
