% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatches.R
\name{.estimateLocalXGradient}
\alias{.estimateLocalXGradient}
\title{Estimate local spatial gradients for scalar X}
\usage{
.estimateLocalXGradient(xy, x, k)
}
\arguments{
\item{xy}{n x 2 coordinate matrix.}

\item{x}{Numeric vector of length n.}

\item{k}{Number of nearest neighbors per cell.}
}
\value{
Matrix (n x 2) of local gradients.
}
\description{
Estimate local spatial gradients for scalar X
}
