% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatches.R
\name{.initKmeans}
\alias{.initKmeans}
\title{Initialize patches with kmeans on xy}
\usage{
.initKmeans(xy, npatches)
}
\arguments{
\item{xy}{n x 2 coordinate matrix.}

\item{npatches}{Number of patches.}
}
\value{
Character vector of patch IDs.
}
\description{
Initialize patches with kmeans on xy
}
