% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.nearestNeighborGraph}
\alias{.nearestNeighborGraph}
\title{Adapted from InSituCor:::nearestNeighborGraph
For each cell identify \code{N} nearest neighbors in Euclidean space and
create an edge between them in graph structure}
\usage{
.nearestNeighborGraph(xy, N)
}
\arguments{
\item{N}{number of nearest neighbors}

\item{x}{spatial coordinate}

\item{y}{spatial coordinate}
}
\value{
sparse adjacency matrix with distances
}
\description{
Edges will only be created for cells that have the same \code{subset} value,
usually the slide column id but could also be a slide plus FOV id to only
create edges within an FOV.
}
