% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.neighbor_tabulate}
\alias{.neighbor_tabulate}
\title{Adapted from InSituCor:::nearestNeighborGraph
for each cell, tabulate the distinct values of x over its neighbors:}
\usage{
.neighbor_tabulate(x, neighbors)
}
\arguments{
\item{x}{A vector of categorical values}

\item{neighbors}{A (probably sparse) adjacency matrix}
}
\description{
Adapted from InSituCor:::nearestNeighborGraph
for each cell, tabulate the distinct values of x over its neighbors:
}
