% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaanalysis.R
\name{.patchMetaSummary}
\alias{.patchMetaSummary}
\title{Compute per-patch summaries of cell metadata}
\usage{
.patchMetaSummary(patch, cellmeta, all_patches)
}
\arguments{
\item{patch}{Patch assignment vector.}

\item{cellmeta}{Cell metadata data frame.}

\item{all_patches}{Character vector of all patch names.}
}
\value{
List of precomputed summary statistics.
}
\description{
Compute per-patch summaries of cell metadata
}
