% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaanalysis.R
\name{.patchNeighborNetwork}
\alias{.patchNeighborNetwork}
\title{Build k-nearest-neighbor network among patches based on W}
\usage{
.patchNeighborNetwork(W, k)
}
\arguments{
\item{W}{Patches x features matrix.}

\item{k}{Number of neighbors.}
}
\value{
Integer matrix (npatches x k) of neighbor indices (column indices into W).
}
\description{
Build k-nearest-neighbor network among patches based on W
}
