% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatches.R
\name{.regularizeCov}
\alias{.regularizeCov}
\title{Regularize a 2x2 covariance matrix to limit elongation}
\usage{
.regularizeCov(S, max_elongation)
}
\arguments{
\item{S}{2x2 covariance matrix.}

\item{max_elongation}{Max ratio of eigenvalues.}
}
\value{
Regularized 2x2 covariance matrix.
}
\description{
Regularize a 2x2 covariance matrix to limit elongation
}
