% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{pearsonResiduals}
\alias{pearsonResiduals}
\title{Convert a raw counts matrix to Pearson residuals}
\usage{
pearsonResiduals(y, tot)
}
\arguments{
\item{y}{Raw counts matrix (cells x genes)}

\item{tot}{Numeric vector of total counts per cell (length = nrow(y))}
}
\value{
Matrix of Pearson residuals (same dimensions as y)
}
\description{
Convert a raw counts matrix to Pearson residuals
}
