% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{plotPatches}
\alias{plotPatches}
\title{Plot patches}
\usage{
plotPatches(
  spe,
  patchesID,
  what = c("both", "assignments", "power"),
  with = c("ggplot", "base")
)
}
\arguments{
\item{spe}{An
[`SpatialExperiment`][SpatialExperiment::SpatialExperiment-class] object.}

\item{patchesID}{A character string specifying the name of the reduced
dimension data frame containing the patch assignments returned by
`getPatches()`. This identifier should be one of the character strings
returned by `reducedDimNames(spe)`, and should be of the form
`patches_<response_cell_type>_<explanatory_cell_type>`.}

\item{what}{A character string specifying what to plot: `assignments` plots
the spatial distribution of the patch assignments to cells, `power` to plot
the predictor sum of squares for each patch, and `both` to plot both of
these things. Default value is `both`.}

\item{with}{A character string specifying which plotting system to use,
`ggplot` (default) or `base`. If `ggplot` is chosen but packages
[`ggplot2`][ggplot2::ggplot2-package] and
[`cowplot`][cowplot::cowplot-package] are not installed, the function will
fall back to `base` plotting and issue a message to the user.}
}
\value{
If `with="ggplot"`, the default value, a `ggplot` object is returned.
If `with="base"`, the function does not return anything, and it produces the
plot.
}
\description{
Visualize the spatial distribution of the patches computed by
`getPatches()` and their statistical power.
}
\examples{


## load a subset of 1k cells of the first sample and slice of the MERFISH
## data available in MerfishData::MouseColonIbdCadinu2024()
library(HDF5Array)
library(SpatialExperiment)

fname <- system.file("extdata", "MerfishData1k", package="SpaceMosaic")
spe <- loadHDF5SummarizedExperiment(fname)

spe <- getPatches(spe, cell_type_column="tier1",
                  response_cell_type="Epithelial",
                  explanatory_cell_type="Immune",
                  mmxpixel=0.000109,
                  npatches=10, n_iters=5)

\dontrun{
plotPatches(spe, patchesID="patches_Epithelial_Immune")
}

}
