% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceMosaic-package.R
\docType{package}
\name{SpaceMosaic-package}
\alias{SpaceMosaic}
\alias{SpaceMosaic-package}
\title{SpaceMosaic: Stratified differential expression analysis on spatial data}
\description{
Performs a differential expression analysis on spatial transcriptomics data stratified on patches defined according to a model of interest.
}
\author{
\strong{Maintainer}: Patrick Danaher \email{patrick.danaher@bruker.com} (\href{https://orcid.org/0000-0002-2844-5883}{ORCID})

Authors:
\itemize{
  \item Patrick Danaher \email{patrick.danaher@bruker.com} (\href{https://orcid.org/0000-0002-2844-5883}{ORCID})
  \item Matteo Calgaro \email{matteo.calgaro@unipd.it} (\href{https://orcid.org/0000-0002-3056-518X}{ORCID})
  \item Robert Castelo \email{robert.castelo@upf.edu} (\href{https://orcid.org/0000-0003-2229-4508}{ORCID})
  \item Pere Moles \email{pere.moles@upf.edu} (\href{https://orcid.org/0009-0009-4408-1897}{ORCID})
  \item Peiying Cai \email{peiying.cai@uzh.ch} (\href{https://orcid.org/0009-0001-9229-2244}{ORCID})
  \item Davide Risso \email{davide.risso@unipd.it} (\href{https://orcid.org/0000-0001-8508-5012}{ORCID})
}

}
\keyword{internal}
