% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{.buildContiguityGraph}
\alias{.buildContiguityGraph}
\title{Build contiguity graph: intersection of Delaunay triangulation and 10-NN}
\usage{
.buildContiguityGraph(xy)
}
\arguments{
\item{xy}{n x 2 coordinate matrix.}
}
\value{
Sparse adjacency matrix (n x n).
}
\description{
Build contiguity graph: intersection of Delaunay triangulation and 10-NN
}
