% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatches.R
\name{getPatches}
\alias{getPatches}
\title{Get patches}
\usage{
getPatches(
  spe,
  cell_type_column,
  response_cell_type,
  explanatory_cell_type,
  mmxpixel = NULL,
  npatches,
  bitesize = 0.1,
  maxradius = 0.5,
  roundness = 0.5,
  initwithhotspots = TRUE,
  n_iters = 25,
  alpha = 0.5,
  effectivezerodist = 0.001,
  plotprogress = FALSE
)
}
\arguments{
\item{spe}{An
[`SpatialExperiment`][SpatialExperiment::SpatialExperiment-class] object.}

\item{cell_type_column}{A character string specifying the column name in
`colData(spe)` that contains the cell type annotations for all cell types.}

\item{response_cell_type}{A character string specifying the response cell
type.}

\item{explanatory_cell_type}{A character string specifying the explanatory
cell type.}

\item{mmxpixel}{Conversion factor from pixels to millimeters to apply to the
spatial coordinates. If NULL (default), no conversion is applied.}

\item{npatches}{A positive integer specifying the number of patches to create.}

\item{bitesize}{How far a patch can expand its borders in any iteration.
Controls the "learning rate". Default value is 0.1.}

\item{maxradius}{Cells beyond this distance from a patch centroid can't be
assigned to it. Used to prevent excessively expansive patches. Default value
is 0.5}

\item{roundness}{Tuning parameter from 0-1 guiding patches' tendency to have
smooth borders. Lower values get better performance (more consistent var(X)
per patch) but produce more uneven borders. Default value is 0.5.}

\item{initwithhotspots}{TRUE or FALSE for whether to initialize patches
around var(X) hotspots (TRUE, default value) or just by clustering cells'
positions (FALSE). Anecdotally, TRUE performs better.}

\item{n_iters}{How many iterations to run. Default value is 25.}

\item{alpha}{Number from 0-1. How aggressively patches with low var(X) will
grab cells from bigger neighboring patches. Default value is 0.5.}

\item{effectivezerodist}{Lower threshold for cells' distance to patch
boundaries. Lower values make it less likely that cells near patch boundaries
will be assigned to the smaller patch, which can help prevent patches from
getting too big. Default value is 0.001.}

\item{plotprogress}{Logical for whether to show progress graphically over
iterations. Default value is FALSE.}
}
\value{
A character vector of patch assignments for each cell of the type
specified by `response_cell_type`. The vector is named with the cell IDs.
}
\description{
Assign cells to "patches" for stratified differential expression analysis.
Suggested tuning parameters to use: 
 1. Control granularity: npatches.
 2. Control learning rate: bizesize, then possibly alpha. 
 3. Patch size and shape: maxradius, roundness
 4. Patch position: initwithhotspots
}
\examples{

library(MerfishData)

spe <- MouseColonIbdCadinu2024()

## subset to sample 1 slice 1
mask <- spe$sample_id == 1 & spe$slice_id == 1
spe_sub <- spe[, mask]

patches <- getPatches(spe, cell_type_column="tier1",
                      response_cell_type="Epithelial",
                      explanatory_cell_type="Immune",
                      mmxpixel=0.000109,
                      npatches=40, n_iters=10)

}
