% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches.R
\name{getPatches}
\alias{getPatches}
\title{Get patches}
\usage{
getPatches(
  spe,
  cell_type_column,
  response_cell_type,
  explanatory_cell_type,
  mmxpixel = NULL,
  npatches,
  bitesize = 0.1,
  maxradius = 0.5,
  roundness = 0.5,
  initwithhotspots = TRUE,
  n_iters = 25,
  alpha = 0.5,
  effectivezerodist = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{spe}{An
[`SpatialExperiment`][SpatialExperiment::SpatialExperiment-class] object.}

\item{cell_type_column}{A character string specifying the column name in
`colData(spe)` that contains the cell type annotations for all cell types.}

\item{response_cell_type}{A character string specifying the response cell
type.}

\item{explanatory_cell_type}{A character string specifying the explanatory
cell type.}

\item{mmxpixel}{Conversion factor from pixels to millimeters to apply to the
spatial coordinates. If NULL (default), no conversion is applied.}

\item{npatches}{A positive integer specifying the number of patches to create.}

\item{bitesize}{How far a patch can expand its borders in any iteration.
Controls the "learning rate". Default value is 0.1.}

\item{maxradius}{Cells beyond this distance from a patch centroid can't be
assigned to it. Used to prevent excessively expansive patches. Default value
is 0.5}

\item{roundness}{Tuning parameter from 0-1 guiding patches' tendency to have
smooth borders. Lower values get better performance (more consistent var(X)
per patch) but produce more uneven borders. Default value is 0.5.}

\item{initwithhotspots}{TRUE or FALSE for whether to initialize patches
around var(X) hotspots (TRUE, default value) or just by clustering cells'
positions (FALSE). Anecdotally, TRUE performs better.}

\item{n_iters}{How many iterations to run. Default value is 25.}

\item{alpha}{Number from 0-1. How aggressively patches with low var(X) will
grab cells from bigger neighboring patches. Default value is 0.5.}

\item{effectivezerodist}{Lower threshold for cells' distance to patch
boundaries. Lower values make it less likely that cells near patch boundaries
will be assigned to the smaller patch, which can help prevent patches from
getting too big. Default value is 0.001.}

\item{verbose}{Show progress on the computations. Default value is TRUE.}
}
\value{
The input
[`SpatialExperiment`][SpatialExperiment::SpatialExperiment-class] object with
a new reduced dimension data frame containing the patch assignments for the
cells annotated to the given cell type response. The reduced dimension data
frame is stored as a reduced dimension under the name
`patches_<response_cell_type>_<explanatory_cell_type>`. This data frame is
in fact a [`DataFrame`][S4Vectors::DataFrame-class] object
with a metadata slot where the patch-level information is stored.
}
\description{
Assign cells to "patches" for stratified differential expression analysis.
Suggested tuning parameters to use: 
 1. Control granularity: npatches.
 2. Control learning rate: bizesize, then possibly alpha. 
 3. Patch size and shape: maxradius, roundness
 4. Patch position: initwithhotspots
}
\examples{

## load a subset of 1k cells of the first sample and slice of the MERFISH
## data available in MerfishData::MouseColonIbdCadinu2024()
suppressPackageStartupMessages({
    library(HDF5Array)
    library(SpatialExperiment)
})

fname <- system.file("extdata", "MerfishData1k", package="SpaceMosaic")
spe <- loadHDF5SummarizedExperiment(fname)

spe <- getPatches(spe, cell_type_column="tier1",
                  response_cell_type="Epithelial",
                  explanatory_cell_type="Immune",
                  mmxpixel=0.000109,
                  npatches=10, n_iters=5)
spe

## a new reduced dimension data frame called "patches_Epithelial_Immune"
## has been added to the input SpatialExperiment object
reducedDimNames(spe)

## using the 'reducedDim()' accessor, we can fetch the patch assignments
patches <- reducedDim(spe, "patches_Epithelial_Immune")
patches

## the patch-level information is in the metadata slot of that data frame
metadata(patches)

}
