% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getPatchPolys}
\alias{getPatchPolys}
\title{Find rough polygon boundaries of patches for visualizations}
\usage{
getPatchPolys(xy, patch)
}
\arguments{
\item{xy}{Cells' xy positions}

\item{patch}{Vector of patch assignments, aligned to the rows of xy}
}
\value{
A named list of polygons, one per patch
}
\description{
Find rough polygon boundaries of patches for visualizations
}
