% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchDE.R
\name{hastyDE}
\alias{hastyDE}
\title{Very fast DE without best practices
Runs simple OLS regression on all genes at once using matrix algebra}
\usage{
hastyDE(y, df)
}
\arguments{
\item{y}{the normalized counts matrix (cells x genes)}

\item{df}{a data frame of the variables to be modeled}
}
\value{
A list with:
  - effect: matrix of effect sizes (genes x predictors)
  - se:     matrix of standard errors
  - p:      matrix of p-values
  - residuals: matrix of residuals (cells x genes)
}
\description{
Very fast DE without best practices
Runs simple OLS regression on all genes at once using matrix algebra
}
