% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{nearestNeighborGraph}
\alias{nearestNeighborGraph}
\title{Create spatial network from N nearest neighbors}
\usage{
nearestNeighborGraph(x, y, N, subset = 1)
}
\arguments{
\item{x}{Spatial coordinate.}

\item{y}{Spatial coordinate.}

\item{N}{Number of nearest neighbors.}

\item{subset}{Same length as x, y. Only cells sharing a subset value become
neighbors.}
}
\value{
Sparse adjacency matrix with distances.
}
\description{
For each cell identify \code{N} nearest neighbors in Euclidean space and
create an edge between them in graph structure.
}
\details{
Edges will only be created for cells that have the same \code{subset} value,
usually the slide column id but could also be a slide plus FOV id to only
create edges within an FOV.
}
