% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{neighbor_colMeans}
\alias{neighbor_colMeans}
\title{Column-wise neighbor means}
\usage{
neighbor_colMeans(x, neighbors)
}
\arguments{
\item{x}{A numeric matrix.}

\item{neighbors}{A sparse adjacency matrix.}
}
\value{
Matrix of the same dimensions as \code{x}.
}
\description{
For each cell, compute the mean of each column of \code{x} across its
neighbors.
}
