% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{patchDE}
\alias{patchDE}
\title{patchDE: run DE over all patches}
\usage{
patchDE(y, df, patch)
}
\arguments{
\item{y}{Expression matrix, cells * genes}

\item{df}{Data frame to be used as DE predictors}

\item{patch}{Vector of patch IDs}
}
\description{
patchDE: run DE over all patches
}
