% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchDE.R
\name{patchDE}
\alias{patchDE}
\title{patchDE: run DE over all patches}
\usage{
patchDE(spe, patchesID, method = "hasty", verbose = FALSE)
}
\arguments{
\item{spe}{An
[`SpatialExperiment`][SpatialExperiment::SpatialExperiment-class] object.}

\item{patchesID}{The name of the reduced dimension in `spe` that contains the patch assignments. 
This should be the same name obtained in `getPatches()`.}

\item{method}{The DE method to use. Currently only "hasty" and "limma" are implemented. Future versions may include additional methods. See details for more information.}

\item{verbose}{If TRUE, print additional information about the DE process for each patch.}
}
\value{
A list of patch-level differential expression results. Each element
  corresponds to a model term and contains a nested list with `pvals`,
  `ests`, and `ses` matrices indexed by gene and patch.
}
\description{
patchDE: run DE over all patches
}
\details{
Available DE methods:
- "hasty": A very fast DE method that performs ordinary least squares regression for each gene and predictor using matrix algebra.
- "limma": A DE method that uses limma's linear model fitting and empirical Bayes moderation.
}
\examples{

## load a subset of 1k cells of the first sample and slice of the MERFISH
## data available in MerfishData::MouseColonIbdCadinu2024()
suppressPackageStartupMessages({
    library(HDF5Array)
    library(SpatialExperiment)
})

fname <- system.file("extdata", "MerfishData1k", package="SpaceMosaic")
spe <- loadHDF5SummarizedExperiment(fname)

spe <- getPatches(spe, cell_type_column="tier1",
                  response_cell_type="Epithelial",
                  explanatory_cell_type="Immune",
                  mmxpixel=0.000109,
                  npatches=10, n_iters=5)
de_results <- patchDE(spe, "patches_Epithelial_Immune")

}
