% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchDE.R
\name{patchDE}
\alias{patchDE}
\title{patchDE: run DE over all patches}
\usage{
patchDE(spe, patchesID)
}
\arguments{
\item{spe}{An
[`SpatialExperiment`][SpatialExperiment::SpatialExperiment-class] object.}

\item{patchesID}{The name of the reduced dimension in `spe` that contains the patch assignments. 
This should be the same name obtained in `getPatches()`.}
}
\value{
A list of patch-level differential expression results. Each element
  corresponds to a model term and contains a nested list with `pvals`,
  `ests`, and `ses` matrices indexed by gene and patch.
}
\description{
patchDE: run DE over all patches
}
\examples{

## load a subset of 1k cells of the first sample and slice of the MERFISH
## data available in MerfishData::MouseColonIbdCadinu2024()
suppressPackageStartupMessages({
    library(HDF5Array)
    library(SpatialExperiment)
})

fname <- system.file("extdata", "MerfishData1k", package="SpaceMosaic")
spe <- loadHDF5SummarizedExperiment(fname)

spe <- getPatches(spe, cell_type_column="tier1",
                  response_cell_type="Epithelial",
                  explanatory_cell_type="Immune",
                  mmxpixel=0.000109,
                  npatches=10, n_iters=5)
de_results <- patchDE(spe, "patches_Epithelial_Immune")

}
