% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceMosaic-package.R
\docType{package}
\name{SpaceMosaic-package}
\alias{SpaceMosaic}
\alias{SpaceMosaic-package}
\title{SpaceMosaic: Spatially-Resolved Differential Expression via Elliptical Patch Partitions}
\description{
Partitions spatial transcriptomics tissue into elliptical patches that maximize within-patch design variable variance, then performs per-patch differential expression with Bayesian meta-analysis and spatial subgroup detection.
}
\author{
\strong{Maintainer}: Patrick Danaher \email{patrick.danaher@bruker.com}

}
\keyword{internal}
